+++ lib/Service/IndexMappingService.php
@@ -38,6 +38,7 @@ use OCA\FullTextSearch\Exceptions\NotIndexableDocumentException;
 use OCA\FullTextSearch\Exceptions\ProviderIsNotCompatibleException;
 use OCA\FullTextSearch_Solr\Exceptions\AccessIsEmptyException;
 use OCA\FullTextSearch_Solr\Exceptions\DataExtractionException;
+Use OCA\FullTextSearch_Solr\Utilities\Utils;
 use OCP\Files\IRootFolder;
 use OCP\FullTextSearch\Model\IndexDocument;
 use OCP\ILogger;
@@ -127,7 +128,7 @@ class IndexMappingService {
 
         // Generate any additional metadata files to be associated with the document.
         $doc = $query->createDocument();
-        $doc->id = $this->generateDocumentIdentifier($document->getProviderId(), $document->getId());
+        $doc->id = Utils::generateDocumentIdentifier($document->getProviderId(), $document->getId());
         $query->setDocument($doc);
 
         // Execute the query
@@ -141,10 +142,6 @@ class IndexMappingService {
             unlink($tmpfname);
         }
 
-    }
-
-    private function generateDocumentIdentifier(string $providerId, string $documentId) {
-        return $providerId . ":" . $documentId;
     }
 
     /**
@@ -191,7 +188,7 @@ class IndexMappingService {
         $update = $client->createUpdate();
 
         // add the delete id and a commit command to the update query
-        $update->addDeleteById($this->generateDocumentIdentifier($providerId, $documentId));
+        $update->addDeleteById(Utils::generateDocumentIdentifier($providerId, $documentId));
         $update->addCommit();
 
         // this executes the query and returns the result